Move playSound to App
