+++ src/components/App.js
@@ -5,10 +5,6 @@ import Board from "./Board";
 
 import "./App.css";
 
-// 246ms long
-import menuMoveSound from "../sounds/menuMove.wav";
-import menuSelectSound from "../sounds/menuSelect.wav";
-
 class App extends React.Component {
   state = {
     muted: true,
@@ -72,6 +68,16 @@ class App extends React.Component {
   componentDidMount() {
     window.addEventListener("keydown", this.handleKeypress);
   }
+
+  playSound = (soundKey, startPoint = 0, volume = 1) => {
+    if (this.state.muted) {
+      return;
+    }
+    const audio = document.querySelector(`[data-sound=${soundKey}]`);
+    audio.currentTime = startPoint;
+    audio.volume = volume;
+    audio.play();
+  };
 
   changeGameActive = (newState) => {
     console.log(`Game active: ${newState}`);
@@ -138,6 +144,7 @@ class App extends React.Component {
         }`}
       >
         <Alert
+          playSound={this.playSound}
           content={this.state.alerts.stageAnnouncement.content}
           shown={this.state.alerts.stageAnnouncement.shown}
           name="stageAnnouncement"
@@ -145,6 +152,7 @@ class App extends React.Component {
           dismissAlert={this.dismissAlert}
         ></Alert>
         <Alert
+          playSound={this.playSound}
           content={this.state.alerts.victory.content}
           shown={this.state.alerts.victory.shown}
           menu={{
@@ -184,6 +192,7 @@ class App extends React.Component {
           dismissAlert={this.dismissAlert}
         ></Alert>
         <Alert
+          playSound={this.playSound}
           name="main-menu"
           content={<h1>Pozo</h1>}
           menu={{
@@ -268,12 +277,14 @@ class App extends React.Component {
           shown={this.state.alerts.mainMenu.shown}
         ></Alert>
         <Alert
+          playSound={this.playSound}
           content={this.state.alert.content}
           shown={this.state.alert.shown}
           autodismiss={this.state.alert.autodismiss}
           dismissAlert={this.dismissAlert}
         ></Alert>
         <Alert
+          playSound={this.playSound}
           name="game-over"
           content={this.state.alerts.gameOver.content}
           menu={{
@@ -312,6 +323,7 @@ class App extends React.Component {
           dismissAlert={this.dismissAlert}
         ></Alert>
         <Alert
+          playSound={this.playSound}
           content={this.state.alerts.instructions.content}
           menu={{
             name: "instructions",
@@ -334,15 +346,15 @@ class App extends React.Component {
           shown={this.state.alerts.instructions.shown}
           dismissAlert={this.dismissAlert}
         ></Alert>
-        <audio data-sound="menuSelect" src={menuSelectSound}></audio>
-        <audio data-sound="menuMove" src={menuMoveSound}></audio>
         <Alert
+          playSound={this.playSound}
           content={this.state.alert.content}
           shown={this.state.alert.shown}
           autodismiss={this.state.alert.autodismiss}
           dismissAlert={this.dismissAlert}
         ></Alert>
         <Board
+          playSound={this.playSound}
           changeGameActive={this.changeGameActive}
           isGameActive={this.state.gameActive}
           showAlert={this.showAlert}
